+++ mayan/apps/smart_settings/apps.py
@@ -2,11 +2,14 @@ from __future__ import unicode_literals
 
 from django.utils.translation import ugettext_lazy as _
 
-from common import MayanAppConfig, menu_setup, menu_object
+from common import MayanAppConfig, menu_secondary, menu_setup, menu_object
 from navigation import SourceColumn
 
 from .classes import Namespace, Setting
-from .links import link_namespace_detail, link_namespace_list
+from .links import (
+    link_namespace_detail, link_namespace_list, link_namespace_root_list,
+    link_setting_edit
+)
 from .widgets import setting_widget
 
 
@@ -37,4 +40,10 @@ class SmartSettingsApp(MayanAppConfig):
         menu_object.bind_links(
             links=(link_namespace_detail,), sources=(Namespace,)
         )
+        menu_object.bind_links(
+            links=(link_setting_edit,), sources=(Setting,)
+        )
+        menu_secondary.bind_links(links=(link_namespace_root_list,))
         menu_setup.bind_links(links=(link_namespace_list,))
+
+        Setting.save_last_known_good()