+++ mayan/apps/smart_settings/urls.py
@@ -2,7 +2,7 @@ from __future__ import unicode_literals
 
 from django.conf.urls import url
 
-from .views import NamespaceDetailView, NamespaceListView
+from .views import NamespaceDetailView, NamespaceListView, SettingEditView
 
 urlpatterns = [
     url(
@@ -12,5 +12,9 @@ urlpatterns = [
     url(
         r'^namespace/(?P<namespace_name>\w+)/$',
         NamespaceDetailView.as_view(), name='namespace_detail'
+    ),
+    url(
+        r'^edit/(?P<setting_global_name>\w+)/$',
+        SettingEditView.as_view(), name='setting_edit_view'
     ),
 ]